Move SQLiteUtils to shared
