+++ collect_app/src/main/java/org/odk/collect/android/database/instances/InstanceDatabaseMigrator.java
@@ -4,7 +4,7 @@ import android.database.sqlite.SQLiteDatabase;
 
 import org.odk.collect.androidshared.sqlite.DatabaseMigrator;
 import org.odk.collect.forms.instances.Instance;
-import org.odk.collect.android.utilities.SQLiteUtils;
+import org.odk.collect.androidshared.sqlite.SQLiteUtils;
 
 import java.util.Arrays;
 import java.util.List;
