+++ collect_app/src/main/java/org/odk/collect/android/database/savepoints/SavepointsDatabaseMigrator.kt
@@ -7,8 +7,8 @@ import org.odk.collect.android.database.savepoints.DatabaseSavepointsColumns.FOR
 import org.odk.collect.android.database.savepoints.DatabaseSavepointsColumns.INSTANCE_DB_ID
 import org.odk.collect.android.database.savepoints.DatabaseSavepointsColumns.INSTANCE_FILE_PATH
 import org.odk.collect.android.database.savepoints.DatabaseSavepointsColumns.SAVEPOINT_FILE_PATH
-import org.odk.collect.android.utilities.SQLiteUtils
 import org.odk.collect.androidshared.sqlite.DatabaseMigrator
+import org.odk.collect.androidshared.sqlite.SQLiteUtils
 
 class SavepointsDatabaseMigrator :
     DatabaseMigrator {
